Simplify Google-Services resolution

+++ packages/app/plugin/src/ios/googleServicesPlist.ts
@@ -7,15 +7,18 @@ import {
 import fs from 'fs';
 import path from 'path';
 
-export const withIosGoogleServicesFile: ConfigPlugin<{ relativePath: string }> = (
-  config,
-  { relativePath },
-) => {
+export const withIosGoogleServicesFile: ConfigPlugin = config => {
   return withXcodeProject(config, config => {
+    if (!config.ios?.googleServicesFile) {
+      throw new Error(
+        'Path to GoogleService-Info.plist is not defined. Please specify the `expo.ios.googleServicesFile` field in app.json.',
+      );
+    }
+
     config.modResults = setGoogleServicesFile({
       projectRoot: config.modRequest.projectRoot,
       project: config.modResults,
-      googleServicesFileRelativePath: relativePath,
+      googleServicesFileRelativePath: config.ios.googleServicesFile,
     });
     return config;
   });
@@ -31,6 +34,12 @@ export function setGoogleServicesFile({
   googleServicesFileRelativePath: string;
 }): XcodeProject {
   const googleServiceFilePath = path.resolve(projectRoot, googleServicesFileRelativePath);
+
+  if (!fs.existsSync(googleServiceFilePath)) {
+    throw new Error(
+      `GoogleService-Info.plist doesn't exist in ${googleServiceFilePath}. Place it there or configure the path in app.json`,
+    );
+  }
 
   fs.copyFileSync(
     googleServiceFilePath,
